Improve speed with transposition table

+++ cpp/2048-web.cpp
@@ -6,7 +6,7 @@ Search search(2, 4.0, 47.0, 3.5, 11.0, 700.0, 270.0);
 #ifdef __cplusplus
 extern "C" {
 #endif
-double EMSCRIPTEN_KEEPALIVE jsWork(row_t row1, row_t row2, row_t row3, row_t row4, int dir) {
+float EMSCRIPTEN_KEEPALIVE jsWork(row_t row1, row_t row2, row_t row3, row_t row4, int dir) {
     return search((board_t(row1) << 48) | (board_t(row2) << 32) | (board_t(row3) << 16) | board_t(row4), dir);
 }
 #ifdef __cplusplus
